requirements.txt
# Core framework dependencies
gradio>=4.0.0
gradio-client>=0.8.0
modal>=0.60.0
llama-index>=0.9.0

# MCP Server dependencies
mcp>=1.0.0
fastmcp>=0.1.0

# LlamaIndex components
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-core>=0.9.0

# AI/ML libraries
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Image processing
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Additional dependencies for document processing
pypdf2>=3.0.0
pdf2image>=1.16.0

# For local development
uvicorn>=0.20.0
fastapi>=0.100.0